ggml: fix q1_0_g128 CPU dot product and add x86 SIMD kernels
Fix float-truncation bug in ggml_vec_dot_q1_0_g128_q8_0: the Q8_0
scale factor d1 was accumulated into an int, silently truncating it
to zero for most blocks and producing garbage output on CPU.

Add AVX2 kernel using the same shuffle/bit-mask expansion as q1_0,
processing one 32-element Q8_0 sub-block per iteration (~4x speedup).

Add AVX-512 hybrid kernel using _mm256_mask_blend_epi8 (AVX-512BW+VL)
to expand weight bits in a single instruction, with _mm256_dpbusd_epi32
(AVX-512VNNI) for the dot product (~5.7x speedup on 1.7B model).

+++ ggml/src/ggml-cpu/quants.c
@@ -177,35 +177,35 @@ void ggml_vec_dot_q1_0_g128_q8_0_generic(int n, float * GGML_RESTRICT s, size_t 
     
     
     float sumf = 0.0;
-    
+
     // Each Q1_0_g128 block has 128 elements, each Q8_0 block has 32 elements
     // So we need 4 Q8_0 blocks per Q1_0_g128 block
     for (int i = 0; i < nb; i++) {
         const float d0 = GGML_FP16_TO_FP32(x[i].d);
-        
-        int sumi = 0;
-        
+
+        float sumi = 0.0f;
+
         // Process 4 Q8_0 blocks (4 * 32 = 128 elements)
         for (int k = 0; k < 4; k++) {
             const float d1 = GGML_FP16_TO_FP32(y[i*4 + k].d);
-            
+
             int sumi_block = 0;
-            
+
             for (int j = 0; j < QK8_0; j++) {
                 const int bit_index = k * QK8_0 + j;
                 const int byte_index = bit_index / 8;
                 const int bit_offset = bit_index % 8;
-                
+
                 // Extract bit: 1 = +1, 0 = -1
                 const int xi = ((x[i].qs[byte_index] >> bit_offset) & 1) ? 1 : -1;
                 const int yi = y[i*4 + k].qs[j];
-                
+
                 sumi_block += xi * yi;
             }
-            
+
             sumi += d1 * sumi_block;
         }
-        
+
         sumf += d0 * sumi;
     }
